requirements.txt
numpy==1.13.3
tensorflow-gpu=1.8.0
Keras==2.2.4
git+git://www.github.com/keras-team/keras-contrib@46fcdb9